+++ src/tests.rs
@@ -36,4 +36,5 @@ fn check_struct_sizes() {
         if cfg!(feature = "no_position") { 8 } else { 16 }
     );
     assert_eq!(size_of::<EvalAltResult>(), 72);
+    assert_eq!(size_of::<NativeCallContext>(), 72);
 }